source/requirements.txt
aiofiles==23.1.0
aiohttp==3.8.5
aiosignal==1.3.1
altair==5.0.1
anyio==3.7.1
async-timeout==4.0.2
attrs==23.1.0
backports.zoneinfo==0.2.1
blinker==1.6.2
Brotli==1.0.9
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.4
dataclasses-json==0.5.9
decorator==5.1.1
duckduckgo-search==3.8.3
exceptiongroup==1.1.2
frozenlist==1.3.3
gitdb==4.0.10
GitPython==3.1.32
google-api-core==2.11.1
google-api-python-client==2.93.0
google-auth==2.22.0
google-auth-httplib2==0.1.0
googleapis-common-protos==1.59.1
greenlet==2.0.2
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==0.17.3
httplib2==0.22.0
httpx==0.24.1
hyperframe==6.0.1
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.0.0
Jinja2==3.1.2
jsonschema==4.18.0
jsonschema-specifications==2023.6.1
langchain==0.0.261
langchainplus-sdk==0.0.20
langsmith==0.0.21
lxml==4.9.3
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.19.0
marshmallow-enum==1.5.1
mdurl==0.1.2
multidict==6.0.4
mypy-extensions==1.0.0
numexpr==2.8.4
numpy==1.24.4
openai==0.27.8
openapi-schema-pydantic==1.2.4
packaging==23.1
pandas==2.0.3
Pillow==9.5.0
pkgutil_resolve_name==1.3.10
protobuf==4.23.4
pyarrow==12.0.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydantic==1.10.11
pydeck==0.8.1b0
Pygments==2.15.1
Pympler==1.0.1
pyparsing==3.1.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
referencing==0.29.1
requests==2.31.0
rich==13.4.2
rpds-py==0.8.10
rsa==4.9
six==1.15.0
smmap==5.0.0
sniffio==1.3.0
socksio==1.0.0
SQLAlchemy==2.0.18
streamlit==1.25.0
streamlit-chat==0.1.1
tenacity==8.2.2
toml==0.10.2
toolz==0.12.0
tornado==6.3.2
tqdm==4.65.0
typing-inspect==0.9.0
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==4.3.1
uritemplate==4.1.1
urllib3==1.26.16
validators==0.20.0
yarl==1.9.2
zipp==3.16.0
